Decentralized the BreadCrumbs modal, it was getting too complicated
