+++ src/components/BibleNavModal/BibleNavModal.scss
@@ -74,122 +74,3 @@
   border-radius: 100%;
   background-color: var(--ion-color-tertiary-container);
 }
-
-// BreadCrumbs Modal
-#bread-crumbs-modal {
-  --height: 98vh;
-  .selected-indicator {
-    display: flex;
-    gap: 4px;
-    flex-direction: column;
-  }
-
-  .selected-text-checkbox {
-    width: 100%;
-  }
-
-  ion-grid {
-    display: flex;
-    flex-direction: column;
-    gap: 12px;
-    height: 75%;
-    justify-content: flex-end;
-    &.full-height {
-      height: 100%;
-    }
-  }
-
-  .chat-container {
-    position: relative;
-    bottom: 0;
-    height: 100%;
-    &.max-height {
-      max-height: calc(100% - 70px);
-    }
-    ion-grid {
-      padding: 0;
-      height: 100%;
-    }
-  }
-
-  .breadcrumbs-header {
-    width: 100%;
-  }
-
-  .messages-container {
-    max-height: 100%;
-    min-height: 20px;
-    overflow-y: auto;
-    margin-top: auto;
-
-    .chat-message {
-      white-space: pre-wrap;
-    }
-  }
-
-  .right-align-text {
-    text-align: right;
-  }
-
-  .right-align-self {
-    justify-content: flex-end;
-  }
-
-  .chat-respond {
-    ion-card {
-      background: var(--ion-background-variant);
-    }
-  }
-
-  .breadcrumbs-suggestions-row {
-    overflow-x: auto;
-    height: fit-content;
-    min-height: 40px;
-
-    .breadcrumbs-suggestions-col {
-      display: flex;
-      flex-wrap: nowrap;
-      width: max-content;
-      height: fit-content;
-      //   overflow-x: auto;
-    }
-  }
-
-  .chat-input-row {
-    background: var(--ion-background-variant);
-    border-radius: 12px;
-    ion-col {
-      border-radius: 12px;
-    }
-    ion-textarea,
-    .textarea-fill-solid.sc-ion-textarea-md-h {
-      background: var(--ion-background-variant);
-      --background: var(--ion-background-variant);
-      border: none;
-      border-radius: 12px;
-      --border-radius: 12px;
-      --border-style: none;
-      --border-width: 50%;
-      --highlight-color-focused: none;
-      // --placeholder-opacity: 1;
-    }
-
-    .textarea-label-placement-floating.sc-ion-textarea-md-h
-      textarea.sc-ion-textarea-md {
-      opacity: 1;
-    }
-
-    ion-col {
-      padding: 0;
-      display: flex;
-      align-items: center;
-      .textarea-send-button {
-        height: 100% !important;
-        margin: 0;
-        --border-radius: 12px;
-        --padding-start: 12px !important;
-        --padding-end: 12px !important;
-      }
-    }
-  }
-}
